[FIRRTL][InferWidths] Start inferring the width of invalid values
This adds support to infer the width of invalid values in the
InferWidths pass.  The width of InvalidValueOps is inferred as a special
case, where it will always be set to the width of the thing it is
connected to.

When an InvalidValue of unknown width is used by an operation that is
not a connect, such as an `AddPrimOp`, then there is no way to infer the
width and we will emit an error.

When a single invalid value with unknown width is used by mutliple
operations, the pass will duplicate the invalid value for each use.
This is so that we can infer a different width depending on the
location, as we don't want to impose an artificial constaint that each
location has the same width. This situation easily arises after running
CSE.

InvalidValues are not being constructed as a part of the constraint
problem, although they could be. Instead, we just update their width
when mapping the rest of the solved widths back into the program.