[FIRRTL][LowerToHW] Don't emit warnings for non-local annotations (#2096)
In FIRRTL unprocessed annotations are dropped in LowerToHW, unless the
`-warn-on-unprocessed-annotations` flag is passed,  which  will create a
a warning for each remaining annotation.

When a non-local annotation goes unhandled, it leaves behind some
`NonLocalAnchor` operations and a few "fake" `circt.nonlocal`
annotations along the instance path. This change makes it so that we
drop `NLA` operations in `LowerToHW` instead of erroring on them. This
also does not emit a warning for the intermediate `circt.nonlocal`
annotations along the NLA path, since we will emit a warning for the
real annotation. This should align with the behaviour of local
annotation.