[FIRRTL] Make presence of `inner_sym` equiv to dont-touch
With inner symbols added to many of the FIRRTL declaration operations,
there now arises a problem where another operation may have a reference
to such a symbol, but an transformation deletes the original
declaration. In the HW dialect, the presence of an inner symbol on an
operation is interpreted as a dont-touch. This commit extends the checks
for dont-touch performed within the FIRRTL dialect (specifically during
folding and IMConstProp) such that they consider the presence of a
`inner_sym` attribute equivalent to a `DontTouchAnnotation`.

At a later point we may even decide to remove `DontTouchAnnotation`s
already in the parser and simply ensure that the targeted operations do
have a `inner_sym`, which would have a similar effect. There are some
possibly weird interactions with `LowerTypes` where the annotation is
easier to maintain than the symbols which likely warrants a separate PR
and discussion.