[Handshake] Introduce `handshake.call`
This commit introduces a new operation `handshake.call` for accessing external modules through the handshake dialect. The operation is intended to model the concept of "communication with the outside world through a latency insensitive interface". Currently, this latency-insentive interface is based on the ESI dialect, which we hopefully can mature into being the standard for everything latency-insensitive in-between different kinds of hardware (or software) interactions.

The operation must reference a `hw.module.extern` or `hw.module` operation. This module is expected to have ESI channel in- and outputs with inner types corresponding to the (bare) types used inside the `handshake` IR.

In FIRRTL lowering, the hw module is rewritten to an external FIRRTL module, which can be instantiated within the lowered `handshake` code. Currently, no modifications are done to the handshake bundles that are generated based on the flat types passed to the `handshake.call` operation - the valid/ready/data signals should map directly to the ESI interface, once the target hw module is lowered.

There are a few todo's for future PRs:
- How do we call external modules with no inputs? in handshake, this is solved by the input `ctrl` operand; we do not (yet) have a definition for "module activation" in ESI.
- How do we ensure that the names of the ESI interface signals agree with what we lower to in HandshakeToFIRRTL?