[SV] Make SVTypes.td includable from other dialects
Similar to the refactoring done in HW, move the `SVDialect` definition
into a separate `SVDialect.td` file, such that other dialects can
selectively include `SVTypes.td` to gain access to the declarations
there. This should help structure the dialects in a fashion that allows
portions to be reused and keep repetition down.