[FIRRTL] Fix local wire/node resets not dominating inferred connections (#2109)
Fix an issue in `InferResets` where using a local wire or node as the
async reset for the module could cause dominance issues when the wire
was declared *after* some of things it would end up getting connected
to. A typical example would be:

    %reg = firrtl.reg %clock
    %localReset = firrtl.wire /*FullAsyncResetAnnotation*/

In this case `InferResets` would decide to hook `%localReset` up to the
register `%reg`, creating a dominance problem.

This commit changes the behavior of `InferResets` such that it moves the
local reset declaration up to before the first thing it gets connected
to, possibly converting a `NodeOp` to a `WireOp` if the node's input
value blocks the move due to it not dominating the target site.

Thanks to @youngar for finding this issue!