[ExportVerilog] Fix output port symbols not interpolating properly (#2115)
Fix an issue in `ExportVerilog` where `hw.exportPort` symbols on output
ports of a module would not be properly interpolated into `sv.verbatim`
operations. This commit fixes the issue by adding the output ports to
the symbol cache, and changing the interpolation of port symbols to rely
on the `getPortVerilogName()` function of the `GlobalNameTable`. This
has the benefit of beign consistent with the corresponding module (name
collisions are resolved), and that no `BlockArgument` is required to get
a hold of the output port name.