[Handshake] Disallow multidimensional and dynamic memories (#2111)
This commit adds some verification constraints to handshake memories, ensuring that only statically sized unidimensional memories are valid.

This should not be seen as a "multidimensional memories are forever disallowed". At its core, this commit adds verification for what is currently unsupported in the FIRRTL lowering. Supporting multidimensional memories(registers) could be handled by lowering to FIRRTL registers instead.

As i see it, it should be the job of a prior pass to lower multidimensional memories and (accesses to them), to unidimensional memories. This is arguably also easier done when we're still in a SSACFG region.