[TableGen] Factor common defs out into common `Types.td` (#2120)
Create a CIRCT-wide `include/circt/Types.td` file that can collect the
common type, attribute, and constraint definitions used throughout the
CIRCT project, which may not be specific to one of the given dialects.
Things that land here would generally be candidates for upstreaming to
MLIR, but it's better to have a common place in CIRCT first where they
can incubate.

While we're at it, get rid of some of the type definitions which have
since appeared in upstream MLIR, and factor some first commonalities out
into `Types.td`. There will be quite a few more in the future.