[ExportVerilog] Legalize declaration op names (#2104)
With the introduction of inner name references, operations such as
`sv.verbatim` can interpolate the name of declarations in other modules,
like wires or instances. This requires that all such operations have
their names properly legalized and stabilized before the parallel
emission of Verilog starts. Currently the names of wires are picked
on-the-fly inside `ModuleEmitter`, which is already too late.

This commit extends the name legalization pass and `GlobalNamesTable` to
also cover declaration ops. This allows such operations to be properly
renamed upfront, with the names being frozen at emission time. The
emission code in `ModuleEmitter` then simply uses that pre-frozen name
for all declarations.