[FIRRTL] Fix type hierarchy
The check for `isa<FIRRTLType>(type)` should return false for the
CHIRRTL types `cmemory` and `cmemoryport`.  The problem is that
`isa<FIRRTLType` was just checking if the dialect of the type was
FIRRTL, and so it would return `true` for the CHIRRTL types. This change
hard-codes the types which are subtypes of `FIRRTLType` into the
`FIRRTLType::classof` function.