[FIRRTL] Add FlowKind op interface
This is an op interface to calculate the flow of operations. When an
operation has flow other than `Source`, it can implement this interface
to customize its flow kind.

This is helpful because it makes flow checking of the FIRRTL dialect
extensible.  This is needed to be able to move the CHIRRTL operations
into their own dialect without having the FIRRTL dialect depend on the
CHIRRTL dialect, creating a circular dependency.