[Handshake] Add interface for specializing port naming (#2151)
This commit introduces a new interface `NamedIOInterface` for handshake operations. This interface is intended to capture the naming of the in- and output operands of a handshake operation. The interface implements a default naming convention equal to what's currently being used (in#, out#). However, operations can override this interface to provide more specialized naming, such as the `MuxOp` example used in the commit.

The interface is used in in HandshakeToFIRRTL through a class `PortNameGenerator`. This class handles port name generation for all operands which may be lowered into handshake modules. If the source operand was a handshake operand (which implements the `NamedIOInterface`) the interface is used to generate port names. Else, default names are generated.