[Handshake] Add support for external memories (#1982)
This commit adds a new operation `handshake.extmemory` to the handshake dialect, for accessing external memories.
The `handshake.extmemory` operation has identical semantics to the `handshake.memory` operation in terms of how `handshake.load` and `handshake.store` interacts with it.

The op is lowered by creating a handshake interface for each referencing load- and store operation that referenced the `handshake.extmemory`. This is equivalent to taking the interface that would be generated for a similar `handshake.memory` and moving that interface to the top-level module. It is then up to the external environment to define the handshake'd access to the external memory.