[FIRRTL] Refactor CircuitNamespace, add ModuleNamespace
Factor the common base of `CircuitNamespace` into a new `NamespaceBase`
class and add a `ModuleNamespace` which captures the inner symbols of
the ports and declarations within a module. Also rename the header they
are in to `Namespace.h`.

Also, since `llvm::StringSet` internalizes the strings added to it and
`insert()` returns an iterator with the allocated key as a `StringRef`,
change the `newName()` function to return a `StringRef` instead of the
copying `std::string`.