[ExportVerilog] Raise precedence of `>>>` aka signed shr (#2160)
This raises the precendence of the signed shift operator from the
`LowestPrecedence` to `Shift`.  It was originally `Shift` precedence,
but was changed in #1681. From the issue, it's not clear to me
what the reason for the change was. The current level of precedence is
mismatched from Verilog precedence and is causing brackets to not be
emitted when needed for the `>>>` arguments.

For example, when using a low precedence operator such as the ternary
mux as an argument to the `>>>`, it should surround the mux in
parenthesis.  Currently it is printing:

```verilog
assign o = $signed($signed(x) >>> p ? 45'h5 : 45'h8)
```
When it should be:
```verilog
assign o = $signed($signed(x) >>> (p ? 45'h5 : 45'h8))
```

Changing the precedence does modify one of the test cases to remove an
extra set of surrounding brackets.