[Handshake] Enforce single-use results and function arguments
This commits adds constraints to the Handshake IR, requiring that results of operations and handshake arguments are single-use.

Pros:
- It is now **not** the job of a lowering pass to check if forks or sinks are missing from the IR. Instead, it is a precondition that the Handshake IR is valid.
- Improves handshake IR by making it more robust (verified) through use of MLIRs verification system.
- The verification can guide those who hand-write handshake code, ensuring that it is valid (...if any).

Cons:
- Cannot hand-write Handshake IR and let forks and sinks be inferred by a pass.

I think this is a fair trade-off. Working with verifying the handshake dialect, I've come to experience quite a few issues which could have been avoided by having proper verification constraints on the IR.