[StaticLogic] Verify the pipeline II attribute when present.
This computes a minimum initiation interval for the pipeline based on
the iter args definitions and uses. It then compares the computed II
against the II attribute that was specified when the pipeline was
created.

The same logic could be used to infer an II when one is not specified,
but that is left as a future enhancement.