[HIR] Banked memrefs instantiate multiple memrefs now.
Earlier a banked memref would instantiate only one memory and hence a
the verilog module that instantiates the memref was specialized for
total number of banks in the memref.
Because of this multiple verilog modules had to be written for different
different number of banks even though the number of elements per bank
was same.
Now the memrefloweringpass instantiates multiple memories, one for each
bank. Thus we only need one verilog module for each unique
(numElementsPerBank, size of element) pair.

Once hw dialect supports parameteric types, we can parameterize the
above as well and then we will have only one verilog module per type of
memory.