[Handshake] Add script for visually debugging a handshake program
This is a script which couples together a .dot representation of a CIRCT
handshake program, and a .vcd trace of a simulation on the same program.
This is done by correlating the .vcd hierarchy with the DOT hierarchy, and
correlating the VCD signals with what looks to be the same values in the DOT
file. Since .dot vertices are created on the handshake operation level,
the script relies on additional attributes in the .dot file to disambiguated
which in- and output edges of a node corresponds to which signal in the .vcd.

.dot files are rendered to .svg images and served by a web server. This server
runs a simple web page which will automatically update the shown image.