[FIRRTL] GCT Interface: Allow top as parent
Update the Grand Central interfaces/views pass to allow for the parent
to be the top module.  Previously there was an incorrect requirement
that the parent module had to be instantiated exactly once.

This is done to enable a pattern where assertions are placed in the test
harness (the top module) and interfaces must therefore work when the
parent is the top module.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>