[FIRRTL] Use inner name refs for GCT interface emission (#2194)
Make use of `InnerRefAttr`s inside the `GrandCentralPass` to refer to
the interface instances and remote declarations targeted by the mapping
module. This makes the names emitted as part of the `sv.verbatim` ops
symbolic and allows `ExportVerilog` to substitute them with the final
values during emission.

Since nested interface instances are not working at the moment, we still
hard-code the name of the interface signals. This is technically only
needed for the case where an interface would contain a nested interface.
However, the way we have to traverse these nested interfaces in the code
makes it very hard to decide upfront whether an interface will be
referred to by name, or by symbol, because the fact whether it was
emitted as a verbatim op (instead of an instance) is only known after
the entire interface has been emitted, which is too late for symbol
substitution.

Some follow-up work should rework the SV interface ops a bit. More
specifically, `sv.interface` has the `SymbolTable` trait, which blocks
resolving the name of nested interfaces to their declaration (because
symbol tables don't resolve into their parent ops). It would be much
more straightforward to have interfaces *not* be symbol tables, and use
`InnerRefAttr` where appropriate. At that point, GCT can properly emit
the nested interfaces and the above problem goes away.