[Transforms] add `memref` flattening pass
This is a pass which flattens statically sized memories and indexing into them by elaborating multi-dimensional indexing into a tree of multipliers and adders. The motivation of this is currently to be able to lower programs having multi-dimensional memories through the handshake flow, which currently only supports single-dimensional memories when lowering to FIRRTL.

This also creates a "Transforms" CIRCT library, following upstream MLIR conventions.