[CombFolds] Generalize the previous pattern significantly.
Handle a wide range of `mux(cond, a, a|b|c)` sorts of patterns, where
an operand to the mux appears in an operand to the subexpr on the
other side.  We handle and, or, xor for now.

This happens a surprising amount for a wide range of operators.  This
simplifies a number of things in CSRFile from RocketCore.fir, for
example, things like:

  _T_170 ? _T_180 | ibuf_io_inst_0_bits_rvc : ibuf_io_inst_0_bits_rvc;
into:
  _T_170 & _T_180 | ibuf_io_inst_0_bits_rvc

and:
  io_fcsr_flags_valid ? reg_fflags | io_fcsr_flags_bits : reg_fflags
into:
  {5{io_fcsr_flags_valid}} & io_fcsr_flags_bits | reg_fflags

and:
   _T_79 ? ~_T_81 & reg_mstatus_mie : reg_mstatus_mie
into:
   (~_T_79 | ~_T_81) & reg_mstatus_mie

This is part of Issue #2095