[FIRRTL] Allow more zero-width folds
Now that upstream `APInt` and friends properly support values of zero
bits, some of the guards against zero width types in `FIRRTLFolds` have
become obsolete. Remove these to generally operate on anything that has
a known width, zero and up. Some operations need to work with values
that are one bit narrower than the input, for example to compute things
like `inputWidth - 1` in signed right shifts. Make these fold to a
constant zero early on, or skip canonicalization altogether.

This follows up on a suggestion by @lattner in #2207.