[CombFolds] Shrink the extract in `extract(thing)&cst` when possible.
If the constant has leading or trailing bits, then we can shrink the
extract (and potentially eliminate the `and` if it is otherwise a
consequtive series of bits.  This has a profound impact on the ALU
module of RocketCore.fir, Turning a mess like this:

 wire [3:0] _T_20 = io_in1[3:0] & 4'hC | _T_13[7:4] & 4'h3;
 = {io_in1[0], io_in1[1:0] & 2'h2 | _T_20[3:2] & 2'h1, _T_20[3:2] & 2’h2, ...

into:

 = {io_in1[0], io_in1[1], io_in1[2], io_in1[3], io_in1[4],

Which can be further reassembled by future rewrites.