[FIRRTL/IMConstantProp] Add new state "ValidValue"
Currently, when we fail to fold operations, we immediately change
their lattice values to be overdefined in the process of fixpoint
iterations. The problem here is that we may success to fold after
a number of iterations. This commit fix the issue by adding new
lattice value called "ValidValue" into between "InvalidValue" and
"Constant".