[CombFolds] Fix issue in icmp ult/ugt extract fold (#2250)
Fix an issue introduced in 3664712 where ult comparisons against a
constant of the shape `0b111..111000..000` would be converted to an
extract operation. The fold is based on the observation that in an icmp
`x < cst`, where `cst` has this shape of a bunch of leading ones
followed by all zeros, all values for `x` for which the comparison would
return a 0 are the values that have all the upper bits set (those that
are a one in the constant). Therefore the comparison is equivalent to
checking if those upper bits are not all ones. The implementation in
3664712 erroneously checked the lower bits of the number.