[ExportVerilog] Remove parallel assert, XFAIL, add comment
Remove the `isInParallelMode` flag and the corresponding assertion,
which are no longer needed since the emission phase of `ExportVerilog`
treats the IR as immutable, which allows for pretty straightforward
parallelization. All mutation is limited to the preparation phase which
lives in the separate `PrepareForEmission.cpp` file.

Also remove the XFAIL on one of the FIRRTL Grand Central integration
tests, which used to trigger this assertion.

Add comments to the `ExportVerilog.cpp` and `PrepareForEmission.cpp`
files explaining the separation into an IR-mutating preparation and
IR-immutable emission phase.