[ExportVerilog] Fix bind port names not being legalized (#2228)
Fix an issue in `ExportVerilog` where the signal names used as port
connections in emitted `bind` statements would not use the legalized
name.

PR #2234 has moved the calls to `prepareHWModule` to a point before the
global name legalization. This now causes the temporary wires that were
introduced by that preparation to have a legalized name, which `bind`
statement emission requires for the port connections.

The remaining work is to adjust the `getNameRemotely` function which
provides the names for the bind port connections to properly see through
the IR pattern used to make output ports of the instance always connect
to a wire:

    %tmpWire = sv.wire
    %output = hw.instance ... () -> (output: i1) {doNotPrint = true}
    sv.assign %tmpWire, %output

This change updates this function to look through any `sv.assign` ops
from instance outputs to wires, and produces that wire's legalized name.

Also adds an assert to ensure that we always have a legal name to use
for the bind port connections. This used to be some optional code that
came up with a name on the spot.