[FIRRTL] Make Annotation hold on to the actual attribute
When an annotation targets only part of the thing it is attached to, it
is represented using a `SubAnnotationAttrs`, which contains the field ID
of its target.

The Annotation wrapper class can be only be created from a
DictionaryAttr.  When using a SubAnnotationAttr, it is created from its
internal DictionaryAttr, and the stored field ID is ignored. This makes
it impractical to use `SubAnnotationAttrs` with most of the Annotation
APIs.

This change makes it so the Annotation wrapper holds on to the original
annotation attribute, instead of just a DictionaryAttr.  The annotation
wrapper will check which kind of annotation it has internally when calls
are made.

This change then adds a call to `getFieldID`, which return 0 for regular
annotations, and the field id for `SubAnnotations`.

This also changes the equality comparison for two annotations: a
Annotation created from a SubAnnotationAttr will no longer be equal to a
regular annotation when the dictionary attribute is the same.