[FIRRTL] Add method to insert ports into instance ops
This adds a method to insert ports into an instance op.  Since this
increases the number of results of the instance op, it actually has to
be cloned.  It is assumed that afterwords, the user will manually call
replaceAllUses and erase the original instance operation.  This
functionality is useful for any pass that adds ports to modules.