[Transforms] Add --flatten-memref-calls pass (#2257)
This pass is intended to be a "software"-like companion to --flatten-memref. Whereas --flatten-memref will flatten all memref operations and function signatures to unidimensional memories and accesses thereto, this pass can be used to transform calls to such flattened operations. In other words, this pass expects the parent function of a `callOp` to be able to support multidimensional memories, but rewrites calls _to_ flattened functions.

This is done by converting memref arguments through a `memref.subview` operation. Any called function which had its signature rewritten will be replaced by a private function definition having a signature of the rewritten function type.