[FIRRTL] Test SFC invalid behaviors, NFC
Add an end-to-end test of firtool that checks for compliance with the
multiple Scala FIRRTL Compiler (SFC) interpretations of invalid.  This
test is added to lock in behavior around invalid as this is known to
cause formal equivalence failures, e.g., if a register has a reset in
CIRCT, but does not in the SFC formal equivalence will rightly complain.
This test can likely be removed or weakened once better semantics for
invalid value are defined other than "whatever the SFC does".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>