[FIRRTL] Canonicalize Invalid Connects to Zero
Extend the behavior of connect canonicalization to convert a connection
to an InvalidValueOp to a connect to zero.  This is done to enforce
compliance with one Scala FIRRTL Compiler (SFC) interpretation of
"invalid is constant zero".

This change requires a RemoveResets pass (#2287) to run before the
canonicalization that this PR adds.  This is because registers require a
separate interpretation of invalid where "invalid is undefined when used
as the initialization value of a register" (and where invalidness is
determined by looking through wires/connects in the current module).
This latter, "register initialization" interpretation needs to run
before invalids are fully removed by the code added in this commit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>