[FIRRTL] Canonicalize Invalid Connect to Zero (#2288)
Extend the behavior of connect canonicalization to convert a connection
to an InvalidValueOp to a connect to zero.  This is done to enforce
compliance with one Scala FIRRTL Compiler (SFC) interpretation of
"invalid is constant zero".

This change requires a RemoveResets pass (#2287) to run before the
canonicalization that this PR adds.  This is because registers require a
separate interpretation of invalid where "invalid is undefined when used
as the initialization value of a register" (and where invalidness is
determined by looking through wires/connects in the current module).
This latter, "register initialization" interpretation needs to run
before invalids are fully removed by the code added in this commit.

Add an end-to-end test of firtool that checks for compliance with the
multiple Scala FIRRTL Compiler (SFC) interpretations of invalid.  This
test is added to lock in behavior around invalid as this is known to
cause formal equivalence failures, e.g., if a register has a reset in
CIRCT, but does not in the SFC formal equivalence will rightly complain.
This test can likely be removed or weakened once better semantics for
invalid value are defined other than "whatever the SFC does".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>