[FIRRTL] Remove fold `mux(p, invalid, x) -> x`
This commit removes the mux canonicalization which optimizes muxes when
one of the branches is invalid to the value of the other branch.

```mlir
mux(p, invalid, x) -> x
```

This optimization is not actually implemented by the SFC compiler, and
we actually need invalids to be treated as 0 when used in a mux.

We do need a similar optimization when lowering `WhenOp`s to muxes
during `ExpandWhens`.  When a wire is invalidated in one branch of a
when op, we optimize away the invalid.  This was previously handled by
creating a mux operation and relying on the folder to perform the
optimization.

A side effect of this change is that in IMCP we no longer want to pass
unknown lattice values as `InvalidAttr` constants to the Mux folder.
Doing this without the folder in-place causes the mux to fail to fold,
which marks the mux operation as overdefined.  Instead, we can just wait
until SCCP knows the values for all the operands to the Mux op before
attempting to fold.