[SV] Make -ignore-read-enable-mem FIRRTL pedantic
Change the behavior of the `-ignore-read-enable-mem` to use the exact
behavior of Scala FIRRTL Compiler default memory lowering.  This will
now use the read enable to gate the update of the latched address.
Previously, the read enable was used to gate the final read.  The net
effect is that the memory will output the last read value continuously.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>