[LowerToHW] Fix reg init via verbatim ops
Work around a bug in Verilog emission with the disallowLocalVariables
option where temporaries are spilled to wires which violates the
ordering of the initial block used for register randomization.  This
manifests as X's showing up in registers after initialization.  This
avoids this entirely by using explicit registers and verbatim operations
using those registers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>