[Handshake] Add transparent FIFO buffer
This commit adds a transparent FIFO buffer to HandshakeToFIRRTL. The implementation is based on the FIFO of dynamatic (https://github.com/lana555/dynamatic/blob/master/components/elastic_components.vhd) but with use of a counter to track the fullness/emptyness of the FIFO (i suspect that there is a bug in the dynamatic implementation, for a 2 slot FIFO, if head == 0, tail == 1, the buffer is considered full).

No tests are added for this, however, i'm currently working with (and verifying) this in CIRCT-HLS in an integration manner. Once things start migrating from CIRCT-HLS to here, coverage of the semantics of the FIRRTL components should be vastly improved.

An upcoming commit will use the InnerFIFO implementation to also implement a non-transparent fifo (as a replacement for the sequential buffers used currently).